pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "printguard"
version = "0.3.1"
description = "Headless MJPEG to MQTT print monitoring service"
authors = [
    { name = "Oliver Bravery", email = "dev@oliverbravery.uk" }
]
readme = "README.md"
requires-python = ">=3.11"
license = "GPL-2.0-only"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Home Automation",
    "Topic :: Internet :: WWW/HTTP",
]

dependencies = [
    "numpy>=2.4,<3",
    "opencv-python-headless>=4.13,<5",
    "onnxruntime>=1.24,<2",
    "pillow>=12.1,<13",
    "paho-mqtt>=2.1,<3",
]

[project.urls]
Homepage = "https://github.com/Meightem/PrintGuard"
Repository = "https://github.com/Meightem/PrintGuard"
Issues = "https://github.com/Meightem/PrintGuard/issues"

[project.optional-dependencies]
dev = [
    "build>=1.2,<2",
    "mypy>=1.11,<2",
    "pip-tools>=7.4,<8",
    "pytest>=8.3,<9",
    "pytest-cov>=6,<7",
    "ruff>=0.11,<1",
]
model-build = [
    "huggingface_hub>=1.7,<2",
]

[project.scripts]
printguard = "printguard.main:run"
printguard-healthcheck = "printguard.health:run_healthcheck"

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]

[tool.black]
line-length = 88

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.11"
files = ["printguard", "tests"]
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = ["cv2", "onnxruntime", "paho.mqtt.client"]
ignore_missing_imports = true

[tool.coverage.run]
branch = true
source = ["printguard"]

[tool.coverage.report]
show_missing = true
skip_covered = false

[tool.setuptools.packages.find]
include = ["printguard", "printguard.*"]
